Added wrapper for theme
